pyproject.toml
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/#declaring-project-metadata 
# https://packaging.python.org/en/latest/tutorials/packaging-projects/

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools]
py-modules = ["pfp"]

[tool.black]
line-length = 100

[project]
name = "pfp"
version = "0.0.1"
authors = [{ name = "Eugenio Chisari", email = "eugenio.chisari@gmail.com" }]
description = ""
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy==1.23.5",
    "spatialmath-python==1.1.9",
    "prompt-toolkit==3.0.36",
    "ipython<=8.17.2",
    "trimesh==4.3.2",
    "open3d==0.18.0",
    "numba<=0.59.1",
    "zarr<=2.17.2",
    "matplotlib<=3.8.4",
    "torch<=2.1.2",
    "torchvision<=0.16.2",
    "einops==0.7.0",
    "diffusers==0.27.2",
    "composer<=0.21.3",
    "hydra-core==1.3.2",
    "wandb<=0.17.3",
    "av==8.1.0",
    "yourdfpy==0.0.56",
    "geomstats[pytorch]==2.7.0",
    "imagecodecs"
]


[project.optional-dependencies]
dev = []
